Compiled and tested against CB 1.4.7-R1.0 Recommended Added ability to disguise as blocks with block ID over 127

+++ src/pgDev/bukkit/DisguiseCraft/packet/DCPacketGenerator.java
@@ -110,9 +110,9 @@ public class DCPacketGenerator {
     	if (d.type.isBlock()) {
     		loc.setY(loc.getY() + 0.5);
     		
-    		Byte blockID = d.getBlockID();
+    		Integer blockID = d.getBlockID();
     		if (blockID != null) {
-    			data = (int) blockID;
+    			data = blockID.intValue();
     			
     			Byte blockData = d.getBlockData();
     			if (blockData != null) {
